requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
PyMySQL
python-dotenv
httpx
dj-database-url
python-decouple
